add cron for delete notification in preview 2 week  and enterprise and slot changes api by kalim
